Added Pulsect, rebalanced moves to acknowledge the fact that you're only going to be fighting one battle at a time

+++ core/src/duffelmon/MonDisplay.java
@@ -144,7 +144,7 @@ public class MonDisplay extends GameObject {
         return customMoveVars[pos];
     }
     
-    public void setMoveVar(int pos, int val) {
+    public void setMoveVar(int pos, double val) {
         customMoveVars[pos] = val;
     }
